float damage

+++ Program.cs
@@ -36,7 +36,7 @@ public class Ship
         float chance = random.Next(1, 11);
         if (chance >= 2)
         {
-            int damage = random.Next(10, 21) + Cannons;
+            float damage = random.Next(10, 21) + Cannons;
             target.Health -= damage;
             target.Health = Math.Max(target.Health, 0);
         }
